requirements.txt
gymnasium==1.0.0
ipython==8.12.3
matplotlib==3.8.4
pandas==2.2.2
pytest==8.4.1
python-dotenv==1.1.1
PyYAML==6.0.1
scipy==1.11.4
seaborn==0.13.2
tabulate==0.9.0
torch==2.7.0+cu118
wandb==0.21.3
yfinance==0.2.61
